extract to mautic actions
