+++ src/components/EmailSignUpCard.js
@@ -14,6 +14,10 @@ import { CardRow } from "./cardCommon"
 import { ReactComponent as EmailAwardsIcon } from "./../resources/mail-awards.svg"
 import { validateEmail } from "./../helpers/validation"
 import { stringify } from "query-string"
+import {
+	extractJSONResponseFromHTML,
+	submitFormToMautic,
+} from "../mauticActions"
 
 const useStyles = makeStyles((theme) => {
 	return {
@@ -70,65 +74,13 @@ export default function EmailSignUp(props) {
 		handleValidationErrors()
 		if (validateEmail(email) && gdpr) {
 			console.log("success")
-			const { formId, returnValue, formName, messenger } = props
-			const data = stringify({
-				"mauticform[email]": email,
-				"mauticform[formId]": formId || "",
-				"mauticform[return]": returnValue || "",
-				"mauticform[formName]": formName || "",
-				"mauticform[messenger]": messenger || true,
-			})
 			setWaiting(true)
 			try {
-				const response = await fetch(
-					`https://mautic.adex.net/form/submit?formId=${formId}`,
-					{
-						method: "POST",
-						body: data,
-						headers: {
-							"Content-Type": "application/x-www-form-urlencoded;charset=UTF-8",
-							"X-Requested-With": "XMLHttpRequest",
-						},
-					}
-				)
-				const utf8Decoder = new TextDecoder("utf-8")
-				const reader = response.body.getReader()
-				let { value: mauticDataResponse } = await reader.read()
-				mauticDataResponse = mauticDataResponse
-					? utf8Decoder.decode(mauticDataResponse)
-					: ""
-				console.log(mauticDataResponse)
-				//TODO: when we have more email forms this should be extracted in helper
-				const regex = /parent.postMessage\("(.+)".+\)/gm
-				const matches = regex.exec(mauticDataResponse)
-				if (matches && matches.length >= 1) {
-					const message = matches[1]
-					let messageCopy = message
-					const asciiRegex = /\\.../gm
-					let m
-					while ((m = asciiRegex.exec(message)) !== null) {
-						// This is necessary to avoid infinite loops with zero-width matches
-						if (m.index === asciiRegex.lastIndex) {
-							asciiRegex.lastIndex++
-						}
-
-						// The result can be accessed through the `m`-variable.
-						/*eslint no-loop-func: "off"*/
-						m.forEach((match, groupIndex) => {
-							const decoded = String.fromCharCode(match.replace("\\", 0))
-							messageCopy = messageCopy.split(match).join(decoded)
-						})
-					}
-					// not able to JSON parse directly so I had to do this above
-					const messageParsed = JSON.parse(messageCopy)
-					console.log(JSON.parse(messageCopy))
-					setMauticState({
-						successSignUp: !!messageParsed.success,
-						...messageParsed,
-					})
-				} else {
-					console.log(`No matches found:`, mauticDataResponse)
-				}
+				const HTMLResponse = await submitFormToMautic(props)
+				const jsonResponse = await extractJSONResponseFromHTML(HTMLResponse)
+				setMauticState({
+					...jsonResponse,
+				})
 			} catch (error) {
 				// If cors is not enabled for address
 				console.error(error)
